requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
clipboard==0.0.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
install==1.3.5
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyperclip==1.8.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
